fix : Redis 락 경쟁조건 고려해서 매서드 옮김2
